Added refresh button to progress bar on app home page (to refresh job delivery progress).
