+++ app/res/layout/home_screen.xml
@@ -62,12 +62,22 @@
         android:layout_height="wrap_content"
         android:visibility="visible"
         >
+        <ImageView
+            android:id="@+id/home_connect_refresh"
+            android:layout_width="32dp"
+            android:layout_height="32dp"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            android:layout_marginTop="10dp"
+            android:src="@drawable/cloud_backup"
+            android:contentDescription="@string/refresh" />
         <ProgressBar
             android:id="@+id/home_connect_prog_bar"
             style="@style/Base.Widget.AppCompat.ProgressBar.Horizontal"
-            android:layout_width="match_parent"
+            android:layout_width="0dp"
             android:layout_height="50dp"
             app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toStartOf="@id/home_connect_refresh"
             app:layout_constraintTop_toTopOf="parent"
             android:layout_marginLeft="@dimen/content_start"
             android:layout_marginRight="@dimen/content_start"
